♻️ refactor: QueryClientProvider 파일 이동 - /application -> /api/core - DefaultPageProps 타입을 /api/core/query-client.tsx 로 이동
