+++ src/api/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./axios";
+export * from "./query-client";
 export * from "./queryKey";
 export * from "./useCoreInfiniteQuery";
 export * from "./useSuspensedQuery";
